new config structure
* Moves config into its own folder with multiple files per purpose to
improve maintainability and readability of config. This separates
config from app as much as possible so that anyone can update config
without dealing with app related stuff.
