+++ src/app/core/services/index.ts
@@ -2,9 +2,7 @@ import { CommonApiService } from './common-api.service';
 import { RecordApiService } from './record-api.service';
 import { HoldingpenApiService } from './holdingpen-api.service';
 import { AppConfigService } from './app-config.service';
-import { CommonConfigsService } from './common-configs.service';
 import { RecordCleanupService } from './record-cleanup.service';
-import { FieldSplitterService } from './field-splitter.service';
 import { DomUtilsService } from './dom-utils.service';
 import { RecordSearchService } from './record-search.service';
 import { SavePreviewModalService } from './save-preview-modal.service';
@@ -14,9 +12,7 @@ export {
   HoldingpenApiService,
   RecordApiService,
   AppConfigService,
-  CommonConfigsService,
   RecordCleanupService,
-  FieldSplitterService,
   DomUtilsService,
   RecordSearchService,
   SavePreviewModalService
@@ -27,9 +23,7 @@ export const CORE_SERVICES = [
   HoldingpenApiService,
   RecordApiService,
   AppConfigService,
-  CommonConfigsService,
   RecordCleanupService,
-  FieldSplitterService,
   DomUtilsService,
   RecordSearchService,
   SavePreviewModalService
