+++ src/app/shared/config/hep/index.ts
@@ -20,30 +20,12 @@
  * as an Intergovernmental Organization or submit itself to any jurisdiction.
  */
 
-import { Component, Input, ChangeDetectionStrategy, OnInit } from '@angular/core';
+import * as _ from 'lodash';
 
-import { REFERENCE_SPLITTER_MAP } from '../../../shared/constants';
+import { coreHep } from './core.config';
+import { thesisHep } from './thesis.config';
 
-@Component({
-  selector: 're-help-modal-button',
-  templateUrl: './help-modal-button.component.html',
-  styleUrls: [
-    './help-modal-button.component.scss'
-  ],
-  changeDetection: ChangeDetectionStrategy.OnPush
-})
-export class HelpModalButtonComponent implements OnInit {
+export const hep = coreHep;
+export const thesis = _.merge(coreHep, thesisHep);
 
-  modChar: string;
-  referenceSplitterMap = REFERENCE_SPLITTER_MAP;
-
-  ngOnInit() {
-    this.modChar = this.getOSModChar();
-  }
-
-  private getOSModChar(): string {
-    let platform = navigator.platform.toLowerCase();
-    let isMac = platform.search('mac') > -1;
-    return isMac ? 'cmd' : 'ctrl';
-  }
-}
+export { onDocumentTypeChange } from './utils';
