+++ src/app/shared/config/hep/thesis.config.ts
@@ -20,30 +20,12 @@
  * as an Intergovernmental Organization or submit itself to any jurisdiction.
  */
 
-import { Component, Input, ChangeDetectionStrategy, OnInit } from '@angular/core';
+import { JsonEditorConfig } from 'ng2-json-editor';
 
-import { REFERENCE_SPLITTER_MAP } from '../../../shared/constants';
-
-@Component({
-  selector: 're-help-modal-button',
-  templateUrl: './help-modal-button.component.html',
-  styleUrls: [
-    './help-modal-button.component.scss'
-  ],
-  changeDetection: ChangeDetectionStrategy.OnPush
-})
-export class HelpModalButtonComponent implements OnInit {
-
-  modChar: string;
-  referenceSplitterMap = REFERENCE_SPLITTER_MAP;
-
-  ngOnInit() {
-    this.modChar = this.getOSModChar();
+export const thesisHep: JsonEditorConfig = {
+  schemaOptions: {
+    alwaysShow: [
+      'thesis_info'
+    ]
   }
-
-  private getOSModChar(): string {
-    let platform = navigator.platform.toLowerCase();
-    let isMac = platform.search('mac') > -1;
-    return isMac ? 'cmd' : 'ctrl';
-  }
-}
+};
