+++ src/app/shared/config/index.ts
@@ -1,0 +1,8 @@
+import { hep, thesis } from './hep';
+
+export const editorConfigs = {
+  hep,
+  thesis,
+};
+
+export * from './api.config';
