+++ src/app/shared/constants/index.ts
@@ -1,1 +1,2 @@
 export { ISO_LANGUAGE_MAP } from './iso-639-1';
+export { REFERENCE_SPLITTER_MAP } from './reference-splitter-map';
