+++ src/typings.d.ts
@@ -3,8 +3,3 @@
 // https://www.typescriptlang.org/docs/handbook/writing-declaration-files.html
 
 declare var System: any;
-
-interface EditorConfig {
-  schemaOptions?: Object;
-  previews?: Array<Object>;
-}
